% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2plots.R
\name{boxPlot}
\alias{boxPlot}
\title{Normalization check box plot}
\usage{
boxPlot(data, samples)
}
\arguments{
\item{data}{data}

\item{samples}{samples}
}
\value{
boxplot
}
\description{
Normalization check box plot
}
