% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countsfile}
\alias{countsfile}
\title{Countsfile}
\format{
## `countsfile` Contains 20 variables and 24500 observations
\describe{
\item{variables}{MeCP2 WT and KO groups}
\item{observations}{gene count levels}
}
}
\usage{
countsfile
}
\value{
dataset
}
\description{
Dataset use to test final functions.
}
\details{
Count gene expression data for testing functions.
}
\keyword{datasets}
