% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{degsfile}
\alias{degsfile}
\title{degsfile}
\format{
## `degsfile` Contains 4 variables and 13717 observations
\describe{
\item{logFC}{log Fold Change levels}
\item{logCPM}{log counts per million}
\item{PValue}{probability under null hypothesis}
\item{FDR}{False Discovery Rate}
}
}
\usage{
degsfile
}
\value{
dataset
}
\description{
Dataset use to test final functions.
}
\details{
Differential Gene Expression (DEGs) data for testing functions.
}
\keyword{datasets}
