% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalRun.R
\name{finalRun}
\alias{finalRun}
\title{Main analysis function}
\usage{
finalRun(
  count.file,
  genotypes,
  degs.file,
  bin.size,
  shift.size,
  title = "MeCP2 KO"
)
}
\arguments{
\item{count.file}{whole cell RNAseq exon counts counts .txt format}

\item{genotypes}{factor WT vs KO}

\item{degs.file}{whole cell RNA seq DEGs .txt file}

\item{bin.size}{bin size integer}

\item{shift.size}{shift size integer}

\item{title}{dataset title}
}
\value{
main analysis plots
}
\description{
Main analysis function
}
