% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gabel.function.R
\name{gabels.plot}
\alias{gabels.plot}
\title{Gabel Function}
\usage{
gabels.plot(
  mat,
  length.type = "Gene",
  comp.between = "",
  y.axis = "Mean Log2 Fold Change"
)
}
\arguments{
\item{mat}{data}

\item{length.type}{Gene}

\item{comp.between}{comp between}

\item{y.axis}{Mean Log2 Fold Change}
}
\value{
Box plot
}
\description{
Gabel's method or moving average method
}
