% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gabel.function.R
\name{gabelsPlot}
\alias{gabelsPlot}
\title{Plot Moving Average for Comparison Between Groups}
\usage{
gabelsPlot(
  mat,
  length.type = "Gene",
  comp.between = "",
  y.axis = "Mean Log2 Fold Change"
)
}
\arguments{
\item{mat}{A dataframe containing the variables for analysis: typically log fold change 
and gene length.}

\item{length.type}{A string describing the type of length being analyzed, usually "Gene".
This parameter is used to label the x-axis of the plot.}

\item{comp.between}{A string that describes the comparison being made, which will be 
used in the plot title to specify the groups being compared (e.g., "KO vs WT").}

\item{y.axis}{A string to be used as the label for the y-axis of the plot, typically 
"Mean Log2 Fold Change".}
}
\value{
A plot object created by the base `plot` function, representing the moving 
average of the specified value across gene lengths.
}
\description{
`gabelsPlot` applies Gabel's method or the moving average method to a given dataset
for the purpose of comparing two groups, typically knockout (KO) versus wild type (WT).
It calculates the moving average of a specified value, such as log2 fold change, across
a range of gene lengths and generates a plot to visualize trends in the data.
}
\examples{
# Generate toy data for log fold change and gene length
set.seed(123) # For reproducibility
logFC.crude <- runif(1000, min=-2, max=2)
gene.length <- sample(2000:1000000, 1000, replace=TRUE)
df <- data.frame(logFC.crude = logFC.crude, gene.length = gene.length)

# Plot the moving average for the generated data
gabelsPlot(mat = df, length.type = "Gene", 
           comp.between = "KO vs WT", 
           y.axis = "Mean Log2 Fold Change")
}
