% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2plots.R
\name{GeneExpLevels}
\alias{GeneExpLevels}
\title{Gene Expression Level in Microarray data}
\usage{
GeneExpLevels(data, gene.id, genotypes)
}
\arguments{
\item{data}{data}

\item{gene.id}{gene ids}

\item{genotypes}{genotypes}
}
\value{
ggplot object with gene expression levels
}
\description{
Gene Expression Level in Microarray data
}
