% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logofMeans.R
\name{log2FCwithingenotypes}
\alias{log2FCwithingenotypes}
\title{Log2FC between genotypes}
\usage{
log2FCwithingenotypes(dat)
}
\arguments{
\item{dat}{data}
}
\value{
log2 fold change between genotypes
}
\description{
Log2FC between genotypes
}
