% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logofMeans.R
\name{logofMeans.between.A.B}
\alias{logofMeans.between.A.B}
\title{Log of means between two samples}
\usage{
logofMeans.between.A.B(dat, A.samples, B.samples)
}
\arguments{
\item{dat}{Differential analysis results from "DESeqCalculation"}

\item{A.samples}{First sample type}

\item{B.samples}{Second sample type}
}
\value{
log2 fold change of means between two given samples
}
\description{
Computes the mean of each sample, fold change, and log fold change.
}
\examples{
\dontrun{
mat <- rbind(matrix(rnorm(600, sd = 0.3), ncol = 6),
matrix(rnorm(600, mean = 1, sd = 0.3), ncol = 6))
colnames(mat) <- c("WT_1","WT_2","WT_3","WT_4","WT_5","WT_6")
logofMeans.between.A.B(dat = mat, A.samples = 1:3,
B.samples = 1:3)
}
}
