% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logofMeans.R
\name{logofMeans.between.ABC}
\alias{logofMeans.between.ABC}
\title{Log of Means between three samples}
\usage{
logofMeans.between.ABC(dat, A.samples, B.samples, C.samples)
}
\arguments{
\item{dat}{a dataframe. Differential analysis results from "DESeqCalculation"}

\item{A.samples}{First sample type}

\item{B.samples}{Second sample type}

\item{C.samples}{Third sample type}
}
\value{
log2 fold change of means between three given samples
}
\description{
Computes the mean of each sample, fold change, and log fold change.
}
\examples{
dat <- matrix(rnorm(600, sd = 0.3), ncol = 9)
dat <- as.data.frame(dat)
logofMeans.between.ABC(dat = dat, A.samples = 1:3, B.samples = 4:6, C.samples = 7:9)


}
