% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logofMeans.R
\name{logofMeansBetweenABC}
\alias{logofMeansBetweenABC}
\title{Calculate Log2 Fold Change Among Three Sets of Samples}
\usage{
logofMeansBetweenABC(dat, A.samples, B.samples, C.samples)
}
\arguments{
\item{dat}{A dataframe or matrix with rows representing genes and columns representing 
samples. It typically contains count data from RNA-seq experiments.}

\item{A.samples}{A vector of column indices or names for the first sample group.}

\item{B.samples}{A vector of column indices or names for the second sample group.}

\item{C.samples}{A vector of column indices or names for the third sample group.}
}
\value{
A modified dataframe that includes additional columns for the mean expression 
of each sample group, the crude fold change, and the log2 fold change based on the 
differences between the sample means. Rows containing `NA` values after computation 
are excluded from the output.
}
\description{
The function computes the average expression values for three sets of samples (A, B, and C) 
and then calculates the fold change and log2 fold change using the differences between 
these averages. This function is useful for RNA-seq data analysis where comparisons are 
made across three different conditions or time points.
}
\examples{
# Simulate RNA-seq count data for 100 genes across 9 samples
set.seed(123) # for reproducibility
dat <- matrix(rnbinom(n=900, mu=100, size=0.5), ncol=9)
colnames(dat) <- c(paste("Sample_A", 1:3, sep=""), 
                   paste("Sample_B", 1:3, sep=""), 
                   paste("Sample_C", 1:3, sep=""))
rownames(dat) <- paste("Gene", 1:100, sep="")
dat <- as.data.frame(dat)

# Calculate log2 fold change among the three sample groups
logFC_results <- logofMeansBetweenABC(dat = dat, A.samples = 1:3, 
                                      B.samples = 4:6, C.samples = 7:9)
head(logFC_results)
}
