% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mCA}
\alias{mCA}
\title{mCA}
\format{
## `mCA` Contains 6 variables and 17916 observations
\describe{
\item{gene.name}{gene name}
\item{mCA}{methylated Cytosine-Adenine}
\item{CA}{Cytosine-Adenine}
\item{mCG}{methylated Cytosine-Guanine}
\item{CG}{Cytosine-Guanine}
\item{gene.length}{gene length in nucleotides}
}
}
\usage{
mCA
}
\value{
dataset
}
\description{
Dataset use to test final functions.
}
\details{
mCA data for testing functions.
}
\keyword{datasets}
