% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2plots.R
\name{MDSplot}
\alias{MDSplot}
\title{MDS Plot}
\usage{
MDSplot(data, genotypes, conditions)
}
\arguments{
\item{data}{data}

\item{genotypes}{genotypes}

\item{conditions}{conditions}
}
\value{
ggplot obuect MDS plot
}
\description{
MDS Plot
}
