% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gabel.function.R
\name{moving.average.function}
\alias{moving.average.function}
\title{Moving Average Function}
\usage{
moving.average.function(
  dat,
  bin.size,
  shift.size,
  length.type,
  comp.between,
  y.axis
)
}
\arguments{
\item{dat}{data}

\item{bin.size}{bin size}

\item{shift.size}{shift size}

\item{length.type}{Gene}

\item{comp.between}{comp between}

\item{y.axis}{Mean Lof2 Fold Change}
}
\value{
ggplot object
}
\description{
Moving Average Function
}
