% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movingAverageAnalysis.R
\name{movingAverageAnalysis}
\alias{movingAverageAnalysis}
\title{Comprehensive Analysis of RNA-seq Data with Moving Average Visualization}
\usage{
movingAverageAnalysis(
  dat,
  genotypes,
  degs.dat,
  bin.size,
  shift.size,
  title = "MeCP2 KO"
)
}
\arguments{
\item{dat}{Path to the whole cell RNAseq exon counts file in .txt format.}

\item{genotypes}{A factor indicating the classification of samples, typically "WT" versus "KO".}

\item{degs.dat}{Path to the whole cell RNAseq differentially expressed genes (DEGs) file in .txt format.}

\item{bin.size}{Integer specifying the bin size for averaging in the overlap plots.}

\item{shift.size}{Integer specifying the step size for the moving average calculation in the overlap plots.}

\item{title}{A character string to be used as the title for the plots, indicating the dataset.}
}
\value{
A list containing the following elements:
\itemize{
  \item{res}{Object containing the results from DESeq2 analysis.}
  \item{overlapPlots}{List of ggplot objects for overlap plots, including combined and individual plots for all genes, edgeR DEGs, and DESeq2 DEGs.}
  \item{scatterPlot1}{Scatter plot for distribution of genes with log2 fold change greater than 0.}
  \item{scatterPlot2}{Scatter plot for distribution of genes with log2 fold change greater than log2(1.2).}
}
}
\description{
Integrates several steps of RNA-seq data analysis for knockout (KO) versus wild type (WT) comparisons
using the DESeq2 package. The function processes exon count data, executes differential expression
analysis, performs k-means clustering, and generates overlap plots showing log2 fold changes
versus gene length. Additionally, scatter plots for gene distribution and other relevant plots
are produced.
}
\examples{
\dontrun{
library(OverlapPlots)
data(countsfile)
data(refseq)
data(degsfile)
genotypes <- factor(c(rep("WT", 10), rep("KO", 10)), levels = c("KO", "WT"))

# Assuming 'countsfile', 'refseq', and 'degsfile' are loaded datasets
analysis_results <- movingAverageAnalysis(dat = countsfile, genotypes = genotypes,
                                         degs.dat = degsfile, bin.size = 60,
                                         shift.size = 6,
                                         title = "KO/WT whole cell dataset")
print(analysis_results$overlapPlots$combined)
}
library(OverlapPlots)

load(file = system.file("extdata","dat-info","mm10_ncbi-refSeqGene_Dec2019.RData", package = "OverlapPlots"))
## KO/WT chromatin dataset (ref: Fig 2F Boxer et al. Mol Cell 2020)
count.file <- readRDS(system.file("extdata","dat","counts","GSE128178_10WT_10MeCP2_KO_whole_cell_RNAseq_exon_counts.rds", package = "OverlapPlots"))
degs.file <- readRDS(system.file("extdata","dat","DEGs","KO-WT_whole-cell_RNA-seq.rds", package = "OverlapPlots"))
dat <- count.file
degs.dat <- degs.file
genotypes <- factor(c(rep("WT", 10), rep("KO", 10)), levels = c("KO", "WT"))
# Set the first column of dat as row names
rownames(dat) <- dat[, 1]
rownames(degs.dat) <- degs.dat[, 1]
# Remove the first column from dat
dat <- dat[, -1]
degs.dat <- degs.dat[, -1]
bin.size <- 60
bin.size <- 60
shift.size <- 6
wholeCell.KO <- movingAverageAnalysis(dat, genotypes,  degs.dat, bin.size, shift.size,
                        title = "KO/WT whole cell dataset")
}
