% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movingAverageAnalysis.R
\name{movingAverageAnalysismCA}
\alias{movingAverageAnalysismCA}
\title{Global Main analysis function for mCA}
\usage{
movingAverageAnalysismCA(
  count.file,
  degs.dat,
  mCA,
  bin.size,
  shift.size,
  title = "MeCP2 KO"
)
}
\arguments{
\item{count.file}{movingAverageAnalysis results}

\item{degs.dat}{whole cell RNA seq DEGs .txt file}

\item{mCA}{mCA data}

\item{bin.size}{bin size integer}

\item{shift.size}{shift size integer}

\item{title}{dataset title}
}
\value{
main analysis plots for mCA
}
\description{
This figure illustrates the comparison between the Log2FC vs the mC (Fig A).
It further looks into the distribution of the long and short genes in each
bin (Fig B and Fig C) and then investigate if contribution of higher mC comes
from long genes as compared to the short genes (Fig D).
}
\examples{
library(OverlapPlots)
data(countsfile)
data(refseq)
data(degsfile)
data(mCA)
dat <- countsfile
degs.dat <- degsfile
genotypes <- factor(c(rep("WT", 10), rep("KO", 10)), levels = c("KO", "WT"))
bin.size <- 60
shift.size <- 6
wholeCell.KO <- movingAverageAnalysis(dat, genotypes,  degs.dat, bin.size, shift.size,
                        title = "KO/WT whole cell dataset")
mCA <- mCA[mCA$CA >= 5,]
mCA <- mCA[mCA$gene.length >= 4500,]
mCA$mCA.CA <- mCA$mCA/mCA$CA
mCA.sub <- mCA[,c(1,6,7)]
mCA.wholeCell.KO <- movingAverageAnalysismCA(wholeCell.KO$res$results, degs.dat, mCA.sub,
                               bin.size, shift.size,
                               title = "KO/WT whole cell dataset")
}
