% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_wrappers.R
\name{overlap_degs_mCA_wrapper}
\alias{overlap_degs_mCA_wrapper}
\title{wrapper around Overlap plots for DEGs mCA}
\usage{
overlap_degs_mCA_wrapper(
  degs.dat,
  count.dat,
  refseq,
  WT1.idx,
  WT2.idx,
  bin.size,
  shift.size,
  methyl.type,
  degs = TRUE
)
}
\arguments{
\item{degs.dat}{DEGs data}

\item{count.dat}{counts data}

\item{refseq}{reference sequence}

\item{WT1.idx}{WT1 index}

\item{WT2.idx}{WT2 index}

\item{bin.size}{bin size}

\item{shift.size}{shift size}

\item{methyl.type}{methylation type}

\item{degs}{DEGs}
}
\value{
Overlay mC plots
}
\description{
wrapper around Overlap plots for DEGs mCA
}
