% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay.gabel.function.R
\name{overlay.gabels.plot}
\alias{overlay.gabels.plot}
\title{Overlay Gabel's plot}
\usage{
overlay.gabels.plot(
  mat,
  bin.size = 200,
  shift.size = 40,
  comp.between1 = "",
  comp.between2 = ""
)
}
\arguments{
\item{mat}{data}

\item{bin.size}{bin size}

\item{shift.size}{shift size}

\item{comp.between1}{comp between}

\item{comp.between2}{comp between}
}
\value{
gabels plot
}
\description{
Overlay Gabel's plot
}
