% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay.mCA.R
\name{overlay.mC}
\alias{overlay.mC}
\title{mC Overlay Plot}
\usage{
overlay.mC(
  mat,
  bin.size = 200,
  shift.size = 40,
  length.type = "Gene",
  methyl.type = "",
  comp.between1 = "",
  comp.between2 = ""
)
}
\arguments{
\item{mat}{dataframe with first column corresponding to comp.mat, second
column to log FC, and third columns to gene length.}

\item{bin.size}{bin size}

\item{shift.size}{shift size}

\item{length.type}{length type}

\item{methyl.type}{methylation type}

\item{comp.between1}{comp between}

\item{comp.between2}{comp between}
}
\value{
Overlay plots mC
}
\description{
mC Overlay Plot
}
