% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay.gabel.function.R
\name{overlay.moving.average.function}
\alias{overlay.moving.average.function}
\title{Overlay Gabel's moving average function}
\usage{
overlay.moving.average.function(
  dat,
  bin.size,
  shift.size,
  comp.between1,
  comp.between2
)
}
\arguments{
\item{dat}{matrix}

\item{bin.size}{bin size}

\item{shift.size}{shift size}

\item{comp.between1}{comp between}

\item{comp.between2}{comp between}
}
\value{
overlay ggplots objects
}
\description{
Overlay Gabel's moving average function
}
