% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay.mCA.R
\name{overlaymC}
\alias{overlaymC}
\title{Overlay Plots for Methylation and Gene Expression Analysis}
\usage{
overlaymC(
  mat,
  bin.size = 200,
  shift.size = 40,
  length.type = "Gene",
  methyl.type = "",
  comp.between1 = "",
  comp.between2 = ""
)
}
\arguments{
\item{mat}{A dataframe with columns corresponding to comparison matrix values (comp.mat),
log fold change (logFC.crude), gene length, methylation values (mCA.CA), and gene names.}

\item{bin.size}{Numeric, the size of the bin for aggregating genes.}

\item{shift.size}{Numeric, the shift size for the moving window across the genes.}

\item{length.type}{Descriptive string of the length type being analyzed, typically "Gene".}

\item{methyl.type}{Descriptive string of the methylation type, typically something like "mCA/CA".}

\item{comp.between1}{Descriptive string of the first condition or state being compared.}

\item{comp.between2}{Descriptive string of the second condition or state being compared.}
}
\value{
A list containing ggplot objects for the overlay plots.
}
\description{
This function generates overlay plots to analyze the relationship between methylation levels 
(mCA/CA) and gene expression across different gene lengths. It uses binning to aggregate data points 
and visualize trends in methylation and expression changes.
}
\examples{
# Generating a toy dataset
set.seed(123) # For reproducibility
a <- runif(1000, min=-2, max=2)
b <- runif(1000, min=-2, max=2)
d <- runif(1000, min=0, max=0.05)
c <- sample(2000:1000000, 1000, replace=TRUE)
genes <- replicate(1000, paste(sample(letters, 8, replace = TRUE), collapse = ""))
df <- data.frame(comp.mat = a, logFC.crude = b, gene.length = c, mCA.CA = d, gene.name = genes)
# Running the overlaymC function
overlay_plots <- overlaymC(mat = df, bin.size = 60, shift.size = 6, methyl.type = "mCA/CA")
# To view the plots, you would call each element from the list, e.g., print(overlay_plots$plot1)
}
