% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2plots.R
\name{PCAplot}
\alias{PCAplot}
\title{PCA plot}
\usage{
PCAplot(data, genotypes, conditions)
}
\arguments{
\item{data}{data}

\item{genotypes}{genotypes}

\item{conditions}{conditions}
}
\value{
ggplot object PCA plot
}
\description{
PCA plot
}
