% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2plots.R
\name{Plot.Scatter}
\alias{Plot.Scatter}
\title{Scatter Plot for DEGs}
\usage{
Plot.Scatter(dat, log2FC, comp.between, pval = 0.05)
}
\arguments{
\item{dat}{data}

\item{log2FC}{log2 fold change}

\item{comp.between}{comp between}

\item{pval}{p-values}
}
\value{
scatter plot
}
\description{
Scatter Plot for DEGs
}
\examples{
\dontrun{Plot.Scatter(dat = degs.dat[,c("gene.name","logFC", "FDR", "gene.length")],
log2FC = log2(1), comp.between = "")}
}
