% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{refseq}
\alias{refseq}
\title{refseq}
\format{
## `refseq` Contains 6 variables and 35976 observations
\describe{
\item{chr}{chromosome number}
\item{tx.start}{start location}
\item{tx.end}{end location}
\item{strand}{positive or negative}
\item{gene.name}{gene name}
\item{gene.length}{gene length in nucleotides}
}
}
\usage{
refseq
}
\value{
dataset
}
\description{
Dataset use to test final functions.
}
\details{
Reference sequence data for testing functions.
}
\keyword{datasets}
