% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay.gabel.function.R
\name{studentt.test2}
\alias{studentt.test2}
\title{p-values from 2 sample t-test}
\usage{
studentt.test2(m1, m2, s1, s2, n1, n2 = n1, m0 = 0, equal.variance = FALSE)
}
\arguments{
\item{m1}{the sample means}

\item{m2}{the sample means}

\item{s1}{the sample standard deviations}

\item{s2}{the sample standard deviations}

\item{n1}{the same sizes}

\item{n2}{the same sizes}

\item{m0}{the null value for the difference in means to be tested for. Default is 0}

\item{equal.variance}{whether or not to assume equal variance. Default is FALSE.}
}
\value{
p-values
}
\description{
p-values from 2 sample t-test
}
