% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_wrappers.R
\name{WTgrp_kmeans}
\alias{WTgrp_kmeans}
\title{K-means original}
\usage{
WTgrp_kmeans(control_mat, centers = 2, iter.max = 1000)
}
\arguments{
\item{control_mat}{a Dataframe containing the control (WT) subset from the results
of Differential Expression analysis generated by "DESeqCalculation"}

\item{centers}{number of centers/clusters default = 2}

\item{iter.max}{max number of iterations default = 1000}
}
\value{
labeled index from k-means clustering
}
\description{
"K-means original" takes control (WT) subset of "DESeqCalculation" results
(which corresponds to Differential Expression analysis). Then, performs k-means
clustering with 2 clusters and outputs corresponding index groups.
}
\examples{
mat <- rbind(matrix(rnorm(1000, sd = 0.3), ncol = 10),
matrix(rnorm(1000, mean = 1, sd = 0.3), ncol = 10))
WTgrp_kmeans(control_mat = mat)
}
