% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeqCalculation.R
\name{DESeqCalculation}
\alias{DESeqCalculation}
\title{DESeqCalculation}
\usage{
DESeqCalculation(dat, genotypes, fc = 1.15)
}
\arguments{
\item{dat}{RNAseq exon counts data. Table with samples in columns and gene in rows.}

\item{genotypes}{factor WT and KO.}

\item{fc}{fold change.}
}
\value{
Differential analysis results and plots.
}
\description{
"DESeqCalculation" takes RNAseq exon counts data and genotypes. Performs
intermediate calculations and results of differential gene expression analysis.
Generates boxplots showing gene expression distributions per sample, PCAplots
separating different KO and WT conditions, differential analysis
gene expression of up and down regulated heatmaps with hierarchical clustering
on samples and genes with euclidean and correlation distances respectively,
MA plot (gene count vs fold change), histogram of p-values with base mean
greater than 1, and returns results.
}
\examples{
# Generate arbitrary toy data
dat <- matrix(rnbinom(n=5000, mu=100, size= 1/0.5),ncol=10)
for (x in 1:250) {
dat[x,] <- dat[x,] + sample(800:1200, 10, replace=TRUE)
}
for (x in 250:500) {
dat[x,] <- dat[x,] + sample(0:100, 10, replace=TRUE)
}
for (x in 1:5) {
dat[,x] <- dat[,x] + sample(800:1000, 5, replace=TRUE)
}
dat <- matrix(as.numeric(dat), ncol = ncol(dat))
colnames(dat) <- c("MeCP2_WT_1",	"MeCP2_WT_2",	"MeCP2_WT_3",
"MeCP2_WT_4",	"MeCP2_WT_5",	"MeCP2_KO_1",	"MeCP2_KO_2",	"MeCP2_KO_3",
"MeCP2_KO_4",	"MeCP2_KO_5")
vec <- as.character(seq(1,500, by = 1))
row.names(dat) <- vec
genotypes <- factor(c(rep("WT", 5), rep("KO", 5)), levels = c("KO", "WT"))

# Run Differential Analysis
DESeqCalculation(dat = dat,genotypes = genotypes, fc = 1.15)

\dontrun{
genotypes <- factor(c(rep("WT", 10), rep("KO", 10)), levels = c("KO", "WT"))
dat <- read.table("../../dat/counts/GSE128178_10WT_10MeCP2_KO_whole_cell_RNAseq_exon_counts.txt",
                  sep = "\t", stringsAsFactors=F, header = TRUE, row.names = 1)
wholeCell.KO <- DESeqCalculation(dat = dat, genotypes = genotypes, fc = 1.15)
}

}
