% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gabel.function.R
\name{gabelsPlot}
\alias{gabelsPlot}
\title{Gabel Function}
\usage{
gabelsPlot(
  mat,
  length.type = "Gene",
  comp.between = "",
  y.axis = "Mean Log2 Fold Change"
)
}
\arguments{
\item{mat}{data}

\item{length.type}{Gene}

\item{comp.between}{comp between}

\item{y.axis}{Mean Log2 Fold Change}
}
\value{
Box plot
}
\description{
Gabel's method or moving average method for KO vs WT
}
\examples{
# generate toy data
b <- runif(1000, min=-2, max=2)
c <- sample(2000:1000000, 1000, replace=TRUE)
df <- data.frame(logFC.crude = b, gene.length = c)
gabelsPlot(mat = df)
}
