% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logofMeans.R
\name{logofMeans.between.A.B}
\alias{logofMeans.between.A.B}
\title{Log of means between two samples}
\usage{
logofMeans.between.A.B(dat, A.samples, B.samples)
}
\arguments{
\item{dat}{Differential analysis results from "DESeqCalculation"}

\item{A.samples}{First sample type}

\item{B.samples}{Second sample type}
}
\value{
log2 fold change of means between two given samples
}
\description{
Computes the mean of each sample, fold change, and log fold change.
}
