% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movingAverageAnalysis.R
\name{movingAverageAnalysis}
\alias{movingAverageAnalysis}
\title{Global Main analysis function}
\usage{
movingAverageAnalysis(
  dat,
  genotypes,
  degs.dat,
  bin.size,
  shift.size,
  title = "MeCP2 KO"
)
}
\arguments{
\item{dat}{whole cell RNAseq exon counts counts .txt format}

\item{genotypes}{factor WT vs KO}

\item{degs.dat}{whole cell RNA seq DEGs .txt file}

\item{bin.size}{bin size integer}

\item{shift.size}{shift size integer}

\item{title}{dataset title}
}
\value{
main analysis plots
}
\description{
This function is intended to generate a figure where it includes the Overlap
plots comparing log2FC vs the gene length. This plot is generated using Boxer
 et al KO/WT whole cell dataset.
}
\examples{
library(OverlapPlots)
data(countsfile)
data(refseq)
data(degsfile)
dat <- countsfile
degs.dat <- degsfile
genotypes <- factor(c(rep("WT", 10), rep("KO", 10)), levels = c("KO", "WT"))
bin.size <- 60
shift.size <- 6
wholeCell.KO <- movingAverageAnalysis(dat, genotypes,  degs.dat, bin.size, shift.size,
                        title = "KO/WT whole cell dataset")
}
