% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movingAverageAnalysis.R
\name{movingAverageAnalysismCA}
\alias{movingAverageAnalysismCA}
\title{Methylation Context Analysis with Differential Gene Expression}
\usage{
movingAverageAnalysismCA(
  count.file,
  degs.dat,
  mCA,
  bin.size,
  shift.size,
  title = "MeCP2 KO"
)
}
\arguments{
\item{count.file}{A dataframe resulting from the `movingAverageAnalysis` containing
normalized count data along with gene names and fold changes.}

\item{degs.dat}{Path to the file with differential gene expression data from RNA-seq
analysis in .txt format.}

\item{mCA}{A dataframe with methylation context data, including gene names, mC values,
and gene lengths.}

\item{bin.size}{Integer specifying the bin size for averaging in the plots.}

\item{shift.size}{Integer specifying the step size for the moving average calculation
in the plots.}

\item{title}{A string indicating the title of the dataset for plot annotations.
Default is "MeCP2 KO".}
}
\value{
A list containing the following components:
\itemize{
  \item{res}{An object with various analysis results and data subsets used in plotting.}
  \item{combined}{A composite ggplot object that includes all the generated plots.}
}
}
\description{
Performs an integrated analysis of methylation context (mCA) data with RNA-seq
differential expression results. It generates comparative plots to examine log2
fold change versus methylation context (mC), distribution of gene lengths within
bins, and the contribution of methylation from long versus short genes.
}
\examples{
\dontrun{
library(OverlapPlots)
# Assuming 'countsfile', 'degsfile', and 'mCA' are loaded datasets
dat <- countsfile # previously loaded or created dataset
degs.dat <- degsfile # previously loaded or created DEGs dataset
genotypes <- factor(c(rep("WT", 10), rep("KO", 10)), levels = c("KO", "WT"))
mCA.sub <- mCA[mCA$CA >= 5 & mCA$gene.length >= 4500,]
mCA.sub$mCA.CA <- mCA.sub$mCA/mCA.sub$CA

# Perform moving average analysis
wholeCell.KO <- movingAverageAnalysis(dat, genotypes, degs.dat, bin.size = 60,
                                      shift.size = 6, title = "KO/WT whole cell dataset")
# Perform mCA analysis
mCA.results <- movingAverageAnalysismCA(wholeCell.KO$res$results, degs.dat, mCA.sub,
                                        bin.size = 60, shift.size = 6,
                                        title = "KO/WT whole cell dataset")
print(mCA.results$combined)
}
library(OverlapPlots)
load(file = system.file("extdata","dat-info","mm10_ncbi-refSeqGene_Dec2019.RData", package = "OverlapPlots"))
## KO/WT chromatin dataset (ref: Fig 2F Boxer et al. Mol Cell 2020)
count.file <- readRDS(system.file("extdata","dat","counts","GSE128178_10WT_10MeCP2_KO_whole_cell_RNAseq_exon_counts.rds", package = "OverlapPlots"))
degs.file <- readRDS(system.file("extdata","dat","DEGs","KO-WT_whole-cell_RNA-seq.rds", package = "OverlapPlots"))
dat <- count.file
degs.dat <- degs.file
mCA <- data.frame(readRDS(system.file("extdata","dat","mCA","CAperGene_10wk_boxer.RDS", package = "OverlapPlots")), stringsAsFactors = FALSE)
genotypes <- factor(c(rep("WT", 10), rep("KO", 10)), levels = c("KO", "WT"))
#' # Set the first column of dat as row names
rownames(dat) <- dat[, 1]
rownames(degs.dat) <- degs.dat[, 1]
# Remove the first column from dat
dat <- dat[, -1]
degs.dat <- degs.dat[, -1]
bin.size <- 60
shift.size <- 6
wholeCell.KO <- movingAverageAnalysis(dat, genotypes,  degs.dat, bin.size, shift.size,
                        title = "KO/WT whole cell dataset")
mCA <- mCA[mCA$CA >= 5,]
mCA <- mCA[mCA$gene.length >= 4500,]
mCA$mCA.CA <- mCA$mCA/mCA$CA
mCA.sub <- mCA[,c(1,6,7)]
mCA.wholeCell.KO <- movingAverageAnalysismCA(wholeCell.KO$res$results, degs.dat, mCA.sub,
                               bin.size, shift.size,
                        title = "KO/WT whole cell dataset")
}
