% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_wrappers.R
\name{overlap_wrapper}
\alias{overlap_wrapper}
\title{wrapper around Overlap plots}
\usage{
overlap_wrapper(
  dat,
  refseq,
  KO.idx,
  WT.idx,
  WT1.idx,
  WT2.idx,
  bin.size,
  shift.size,
  shrink_lfc = FALSE
)
}
\arguments{
\item{dat}{Differential analysis results from "DESeqCalculation"}

\item{refseq}{reference sequence}

\item{KO.idx}{Treatment group index}

\item{WT.idx}{Control group index}

\item{WT1.idx}{WT1 index}

\item{WT2.idx}{WT2 index}

\item{bin.size}{bin size}

\item{shift.size}{shift size}

\item{shrink_lfc}{Shrink log2 fold changes}
}
\value{
results and plots
}
\description{
"Wrapper around overlap plots" takes "DESeqCalculation" results,
k-mean clustering index groups, and computes running average
plots (log2 fold change vs mean gene length).

Uses Overlay Gabel's moving average function which calculates the mean of
log2 fold change and genomic lengths. Computes p-values using student t-test
function, and returns overlay line plots.
}
\examples{
## Load annotation dataset
\dontrun{
load(file = "../../dat-info/mm10_ncbi-refSeqGene_Dec2019.RData")
genotypes <- factor(c(rep("WT", 10), rep("KO", 10)), levels = c("KO", "WT"))
## Load gene counts
dat <- read.table(paste("../../dat/counts/GSE128178_10WT_",
"10MeCP2_KO_whole_cell_RNAseq_exon_counts.txt"),
           sep = "\t", stringsAsFactors=FALSE, header = TRUE, row.names = 1)
# Run DESeq analysis
wholeCell.KO <- DESeqCalculation(dat = dat, genotypes = genotypes, fc = 1.15)
mat <- wholeCell.KO$results[,c(8:27,1,3)]
colnames(mat)[21] <- "gene.name"
## same cluster sizes
grp.idx <- WTgrp_kmeans(control_mat = mat[,1:10])
## Generate overlap plots
res1 <- overlap_wrapper(dat = mat, refseq = refseq, KO.idx = c(11:20),
                        WT.idx = c(1:10), WT1.idx = grp.idx$WT.idx1,
                        WT2.idx = grp.idx$WT.idx2, bin.size = 200,
                        shift.size = 40)
res1$plot
}
}
