% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay.gabel.function.R
\name{overlayGabelsPlot}
\alias{overlayGabelsPlot}
\title{Overlay Gabel's plot}
\usage{
overlayGabelsPlot(
  mat,
  bin.size = 200,
  shift.size = 40,
  comp.between1 = "",
  comp.between2 = "",
  confidenceinterval = 0.5
)
}
\arguments{
\item{mat}{dataframe with first column corresponding to comp.mat, second
column to log FC, and third columns to gene length.}

\item{bin.size}{bin size}

\item{shift.size}{shift size}

\item{comp.between1}{comp between}

\item{comp.between2}{comp between}

\item{confidenceinterval}{Confidence interval to be displayed on plots}
}
\value{
gabels plot
}
\description{
Overlay Gabel's plot
}
\examples{
# generate toy data
a <- runif(1000, min=-2, max=2)
b <- runif(1000, min=-2, max=2)
c <- sample(2000:1000000, 1000, replace=TRUE)
df <- data.frame(comp.mat = a, logFC.crude = b, gene.length = c)
overlayGabelsPlot(mat = df,comp.between1 = "(WT/WT)",
comp.between2 = "(KO/WT)",bin.size = 200,
shift.size = 40, confidenceinterval=0.50)
}
