% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay.mCA.R
\name{overlaymC}
\alias{overlaymC}
\title{mC Overlay Plot}
\usage{
overlaymC(
  mat,
  bin.size = 200,
  shift.size = 40,
  length.type = "Gene",
  methyl.type = "",
  comp.between1 = "",
  comp.between2 = ""
)
}
\arguments{
\item{mat}{dataframe with first column corresponding to comp.mat, second
column to log FC, and third columns to gene length, the forth column contains
the mCA.CA values, and last columns the gene names.}

\item{bin.size}{bin size}

\item{shift.size}{shift size}

\item{length.type}{length type}

\item{methyl.type}{methylation type}

\item{comp.between1}{comp between}

\item{comp.between2}{comp between}
}
\value{
Overlay plots mC
}
\description{
mC Overlay Plot
}
\examples{
# generate toy data
a <- runif(1000, min=-2, max=2)
b <- runif(1000, min=-2, max=2)
d <- runif(1000, min=0, max=0.05)
c <- sample(2000:1000000, 1000, replace=TRUE)
genes <- OpenRepGrid::randomWords(1000)
df <- data.frame(comp.mat = a, logFC.crude = b, gene.length = c, mCA.CA = d,
gene.name = genes)
overlaymC(mat = df,bin.size = 60, shift.size = 6, methyl.type = "mCA/CA")
}
